Extract HarnessAdapter seam: new agent/session_runner/harness/ package
Move argv/env assembly + stream-json parsing (get_response_via_harness,
_run_harness_subprocess, verify_harness_health, build_harness_turn_input,
sentinels) out of agent/sdk_client.py into agent/session_runner/harness/claude.py,
byte-identical. Add base.py (HarnessAdapter protocol, TurnRequest/TurnResult/
TurnEvent) and events.py (normalized event vocabulary). ClaudeHarnessAdapter
wraps get_response_via_harness, translating its callback side channel into a
TurnResult with a normalized events list, emitting session.started first
(Race 1). Additive on_usage callback on get_response_via_harness feeds
TurnResult.usage.

agent/sdk_client.py re-exports the moved names (explicit `as`-aliased idiom,
mid-file to avoid a circular import through agent.session_runner's package
init) so existing non-runner callers (session_completion.py, agent/__init__.py)
are unaffected.

Repoint tests that mocked agent.sdk_client._run_harness_subprocess /
_DISABLE_THINKING_SENTINEL at their new home in
agent.session_runner.harness.claude (mocking the re-exported name no longer
affects the real call site once the implementation moved).

Part of #2000 Task 2.2.
